+++ src/typescript/wastelandVentures.ts
@@ -1,6 +1,7 @@
 import WvActor from "./actor/wvActor.js";
 import WvActorSheet from "./actor/wvActorSheet.js";
 import { CONSTANTS } from "./constants.js";
+import { WvCombat } from "./foundryOverrides/wvCombat.js";
 import { DragRuler } from "./integrations/dragRuler/dragRuler.js";
 import { createWvSpeedProvider } from "./integrations/dragRuler/wvSpeedProvider.js";
 import WvItem from "./item/wvItem.js";
@@ -11,6 +12,9 @@ Hooks.once("init", () => {
   // Register our own Entity classes.
   CONFIG.Actor.entityClass = WvActor;
   CONFIG.Item.entityClass = WvItem;
+
+  // Register our override classes.
+  CONFIG.Combat.entityClass = WvCombat;
 
   registerSystemSettings();